+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/Exporter.java
@@ -29,11 +29,11 @@ import com.microsoft.applicationinsights.agent.internal.common.OperationLogger;
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.ContextTagKeys;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.SeverityLevel;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.AbstractTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.ExceptionTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.MessageTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RemoteDependencyTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RequestTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.AbstractTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.ExceptionTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MessageTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RemoteDependencyTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RequestTelemetryBuilder;
 import com.microsoft.applicationinsights.agent.internal.telemetry.FormattedDuration;
 import com.microsoft.applicationinsights.agent.internal.telemetry.FormattedTime;
 import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryClient;
