+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/models/builders/AbstractTelemetryBuilder.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.exporter.models2;
+package com.microsoft.applicationinsights.agent.internal.exporter.models.builders;
 
 import static com.microsoft.applicationinsights.agent.internal.common.TelemetryTruncation.truncatePropertyValue;
