+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/models/builders/MessageTelemetryBuilder.java
@@ -19,66 +19,37 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.exporter.models2;
+package com.microsoft.applicationinsights.agent.internal.exporter.models.builders;
 
 import static com.microsoft.applicationinsights.agent.internal.common.TelemetryTruncation.truncateTelemetry;
 
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.RemoteDependencyData;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.MessageData;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.SeverityLevel;
 import com.microsoft.applicationinsights.agent.internal.exporter.utils.SanitizationHelper;
 import java.util.HashMap;
 import java.util.Map;
 
-public final class RemoteDependencyTelemetryBuilder extends AbstractTelemetryBuilder {
+public final class MessageTelemetryBuilder extends AbstractTelemetryBuilder {
 
-  private static final int MAX_DATA_LENGTH = 8192;
-  private static final int MAX_RESULT_CODE_LENGTH = 1024;
-  private static final int MAX_DEPENDENCY_TYPE_LENGTH = 1024;
-  private static final int MAX_TARGET_NAME_LENGTH = 1024;
+  private final MessageData data;
 
-  private final RemoteDependencyData data;
-
-  public static RemoteDependencyTelemetryBuilder create() {
-    return new RemoteDependencyTelemetryBuilder(new RemoteDependencyData());
+  public static MessageTelemetryBuilder create() {
+    return new MessageTelemetryBuilder(new MessageData());
   }
 
-  private RemoteDependencyTelemetryBuilder(RemoteDependencyData data) {
-    super(data, "RemoteDependency", "RemoteDependencyData");
+  private MessageTelemetryBuilder(MessageData data) {
+    super(data, "Message", "MessageData");
     this.data = data;
   }
 
-  public void setId(String id) {
-    data.setId(truncateTelemetry(id, SanitizationHelper.MAX_ID_LENGTH, "RemoteDependency.id"));
+  public void setMessage(String message) {
+    data.setMessage(
+        truncateTelemetry(message, SanitizationHelper.MAX_MESSAGE_LENGTH, "Message.message"));
   }
 
-  public void setName(String name) {
-    data.setName(
-        truncateTelemetry(name, SanitizationHelper.MAX_NAME_LENGTH, "RemoteDependency.name"));
-  }
-
-  public void setResultCode(String resultCode) {
-    data.setResultCode(
-        truncateTelemetry(resultCode, MAX_RESULT_CODE_LENGTH, "RemoteDependency.resultCode"));
-  }
-
-  public void setData(String data) {
-    this.data.setData(truncateTelemetry(data, MAX_DATA_LENGTH, "RemoteDependency.data"));
-  }
-
-  public void setType(String type) {
-    data.setType(truncateTelemetry(type, MAX_DEPENDENCY_TYPE_LENGTH, "RemoteDependency.type"));
-  }
-
-  public void setTarget(String target) {
-    data.setTarget(truncateTelemetry(target, MAX_TARGET_NAME_LENGTH, "RemoteDependency.target"));
-  }
-
-  public void setDuration(String duration) {
-    data.setDuration(duration);
-  }
-
-  public void setSuccess(Boolean success) {
-    data.setSuccess(success);
+  public void setSeverityLevel(SeverityLevel severityLevel) {
+    data.setSeverityLevel(severityLevel);
   }
 
   public void addMeasurement(String key, Double value) {
