+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/profiler/GcEventMonitor.java
@@ -22,7 +22,7 @@
 package com.microsoft.applicationinsights.agent.internal.profiler;
 
 import com.microsoft.applicationinsights.agent.internal.configuration.GcReportingLevel;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.EventTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.EventTelemetryBuilder;
 import com.microsoft.applicationinsights.agent.internal.telemetry.FormattedTime;
 import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryClient;
 import com.microsoft.applicationinsights.alerting.AlertingSubsystem;
