+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/telemetry/TelemetryClient.java
@@ -31,14 +31,14 @@ import com.microsoft.applicationinsights.agent.internal.exporter.models.MetricDa
 import com.microsoft.applicationinsights.agent.internal.exporter.models.MetricsData;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.MonitorDomain;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.TelemetryItem;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.AbstractTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.EventTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.ExceptionTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.MessageTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.MetricTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.PageViewTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RemoteDependencyTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RequestTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.AbstractTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.EventTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.ExceptionTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MessageTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MetricTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.PageViewTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RemoteDependencyTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RequestTelemetryBuilder;
 import com.microsoft.applicationinsights.agent.internal.localstorage.LocalFileCache;
 import com.microsoft.applicationinsights.agent.internal.localstorage.LocalFileLoader;
 import com.microsoft.applicationinsights.agent.internal.localstorage.LocalFileSender;
