+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/quickpulse/QuickPulseTestBase.java
@@ -37,9 +37,9 @@ import com.azure.core.test.TestMode;
 import com.azure.core.util.FluxUtil;
 import com.azure.identity.ClientSecretCredentialBuilder;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.TelemetryItem;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.ExceptionTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RemoteDependencyTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models2.RequestTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.ExceptionTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RemoteDependencyTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RequestTelemetryBuilder;
 import com.microsoft.applicationinsights.agent.internal.telemetry.FormattedDuration;
 import com.microsoft.applicationinsights.agent.internal.telemetry.FormattedTime;
 import java.nio.charset.StandardCharsets;
